backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
web3==7.5.0
pyevmasm==0.2.3
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
dune-client==1.7.4
mangum==0.19.0
